DO NOT MERGE Change permissions review mode check.
In ag/1835531 we introduced a check for permissions review mode,
that relies on an overlay resource. Since we can't use overlay
resources for the Unified Build, this CL temporarily introduces
the checking for a build property as well. This possibly won't work
for O/G, but will work for the Feldspar release of the Unified build.

BUG: 35028846
Change-Id: I4c82f242bc680028ecd872589552534a419f0152