Improve preferences for PreferenceDataStore.
I have done some important cleanup of java docs where it wasn't clear
enough that there is a PreferenceDataStore option.

I have fixed incorrect behavior when setting default value while data
store is assigned. Previsouly we looked into SharedPreferences which is
incorrect.

I have also changed behavior of methods getSharedPreferences and
getEditor. These methods now return null when custom data store is
assigned. The reason is that we definetly don't want to use
SharedPreferences if data store is assigned. So the original behavior
was a bit ambiguous. Also if we didn't enforce it, we would have
troubles to change that in the future.

Test: Tests are under the same topic.
Change-Id: Ib7303bae13648767c519edba74ef9a7143e703df