Correctly check expired jobs against active ones
If a job was run before its deadline, and then the deadline landed
while the job was still running, b/c of the 'deadline trumps all'
it would be considered eligible to run again as soon as the previous
iteration completed.

Fix this by checking against the state of the active queue.

Bug 36253320
Test: CTS: job scheduler TimingConstraintsTest

Change-Id: I4f0c1012091bf88776cfe725d7b59f6dcf602c02